feat: list of participants
