+++ en.lproj/Localizable.strings
@@ -137,6 +137,10 @@ We can't get your challenges now, try again";
     "Voltar para meus desafios" = "Back to my challenges";
     "Começar desafio!" = "Start challenge!";
     "Apagar desafio" = "Delete challenge";
+    "Tipo" = "Type";
+    "Convidar Participante" = "Invite Participants";
+    "Ver todos" = "See all";
+    "Detalhes do desafio" = "Challenge details";
 
 // MARK: - Quick - Details - DetailsAlertCases
     "Apagar desafio" = "Delete challenge";
