+++ pt-BR.lproj/Localizable.strings
@@ -141,6 +141,10 @@
     "Desafio finalizado" = "Desafio finalizado";
     "Voltar para meus desafios" = "Voltar para meus desafios";
     "Apagar desafio" = "Apagar desafio";
+    "Tipo" = "Tipo";
+    "Convidar Participante" = "Convidar Participante";
+    "Ver todos" = "Ver todos";
+    "Detalhes do desafio" = "Detalhes do desafio";
 
 // MARK: - Quick - Details - DetailsAlertCases
     "Apagar desafio" = "Apagar desafio";
